+++ app/controllers/clearance/users_controller.rb
@@ -27,14 +27,6 @@ class Clearance::UsersController < Clearance::BaseController
 
   private
 
-  def avoid_sign_in
-    warn "[DEPRECATION] Clearance's `avoid_sign_in` before_filter is " +
-      "deprecated. Use `redirect_signed_in_users` instead. " +
-      "Be sure to update any instances of `skip_before_filter :avoid_sign_in`" +
-      " or `skip_before_action :avoid_sign_in` as well"
-    redirect_signed_in_users
-  end
-
   def redirect_signed_in_users
     if signed_in?
       redirect_to Clearance.configuration.redirect_url